add invoice_items table

+++ database/migrations/2023_08_06_201724_add_column_item_id_in_invoice_item_table.php
@@ -14,7 +14,7 @@ class AddColumnItemIdInInvoiceItemTable extends Migration
     public function up()
     {
         Schema::table('invoice_items', function (Blueprint $table) {
-            $table->unsignedBigInteger('item_id')->index()->comment('id of items table')->after('invoice_id');
+            $table->Integer('item_id')->index()->comment('id of items table')->after('invoice_id');
             $table->foreign('item_id')->references('id')->on('items');
         });
     }
